+++ src/test/common/control/open-control.ts
@@ -3,7 +3,7 @@ import { CommonContext } from "../../../context-types";
 export interface OpenControl<X> {
   //test control
   openMockApp(appName?: string, appId?: string, contextType?: string, targetAppAsString?: boolean, malformedContext?: boolean): void;
-  closeAppWindows(testId: string): Promise<void>;
+  closeMockApp(testId: string): Promise<void>;
 
   //listening
   contextReceiver(contextType: string, expectNotToReceiveContext?: boolean): Promise<X>;
