+++ src/test/v1.2/advanced/fdc3.raiseIntent.ts
@@ -7,6 +7,7 @@ import { APIDocumentation1_2 } from "../apiDocuments-1.2";
 import { ContextType, Intent } from "../support/intent-support-1.2";
 import { IntentApp } from "../../v2.0/support/intent-support-2.0";
 import { AppControlContext } from "../../../context-types";
+import { waitForContext } from "../fdc3-1_2-utils";
 
 declare let fdc3: DesktopAgent;
 const raiseIntentDocs = "\r\nDocumentation: " + APIDocumentation1_2.raiseIntent + "\r\nCause";
@@ -120,57 +121,3 @@ async function closeIntentAppsWindows(testId) {
   await waitForContext("windowClosed", testId, appControlChannel);
   await wait(constants.WindowCloseWaitTime);
 }
-
-const waitForContext = (contextType: string, testId: string, channel?: Channel): Promise<Context> => {
-  let executionListener: Listener;
-  return new Promise<Context>(async (resolve) => {
-    console.log(Date.now() + ` Waiting for type: "${contextType}", on channel: "${channel.id}" in test: "${testId}"`);
-
-    const handler = (context: AppControlContext) => {
-      if (testId) {
-        if (testId == context.testId) {
-          console.log(Date.now() + ` Received ${contextType} for test: ${testId}`);
-          resolve(context);
-          if (executionListener) executionListener.unsubscribe();
-        } else {
-          console.warn(Date.now() + ` Ignoring "${contextType}" context due to mismatched testId (expected: "${testId}", got "${context.testId}")`);
-        }
-      } else {
-        console.log(Date.now() + ` Received (without testId) "${contextType}" for test: "${testId}"`);
-        resolve(context);
-        if (executionListener) executionListener.unsubscribe();
-      }
-    };
-
-    if (channel === undefined) {
-      executionListener = fdc3.addContextListener(contextType, handler);
-    } else {
-      executionListener = channel.addContextListener(contextType, handler);
-      //App channels do not auto-broadcast current context when you start listening, so retrieve current context to avoid races
-      const ccHandler = async (context: AppControlContext) => {
-        if (context) {
-          if (testId) {
-            if (testId == context?.testId && context?.type == contextType) {
-              console.log(Date.now() + ` Received "${contextType}" (from current context) for test: "${testId}"`);
-              if (executionListener) executionListener.unsubscribe();
-              resolve(context);
-            } //do not warn as it will be ignoring mismatches which will be common
-            else {
-              console.log(
-                Date.now() +
-                  ` CHecking for current context of type "${contextType}" for test: "${testId}" Current context did ${context ? "" : "NOT "} exist, 
-had testId: "${context?.testId}" (${testId == context?.testId ? "did match" : "did NOT match"}) 
-and type "${context?.type}" (${context?.type == contextType ? "did match" : "did NOT match"})`
-              );
-            }
-          } else {
-            console.log(Date.now() + ` Received "${contextType}" (from current context) for an unspecified test`);
-            if (executionListener) executionListener.unsubscribe();
-            resolve(context);
-          }
-        }
-      };
-      channel.getCurrentContext().then(ccHandler);
-    }
-  });
-};
