+++ src/test/v1.2/support/open-support-1.2.ts
@@ -6,6 +6,7 @@ import { sleep, wait } from "../../../utils";
 import { AppControlContext } from "../../../context-types";
 import { OpenControl } from "../../common/control/open-control";
 import { APIDocumentation1_2 } from "../apiDocuments-1.2";
+import { closeMockAppWindow, waitForContext } from "../fdc3-1_2-utils";
 
 declare let fdc3: DesktopAgent;
 const openDocs = "\r\nDocumentation: " + APIDocumentation1_2.open + "\r\nCause:";
@@ -62,18 +63,15 @@ export class OpenControl1_2 implements OpenControl<Context> {
     }
   };
 
+  async closeMockApp(testId: string) {
+    await closeMockAppWindow(testId);
+  }
+
   addListenerAndFailIfReceived = async () => {
     const appControlChannel = await fdc3.getOrCreateChannel(constants.ControlChannel);
     await appControlChannel.addContextListener("context-received", (context: AppControlContext) => {
       assert.fail(context.errorMessage);
     });
-  };
-
-  closeAppWindows = async (testId) => {
-    await broadcastCloseWindow(testId);
-    const appControlChannel = await fdc3.getOrCreateChannel(constants.ControlChannel);
-    await waitForContext("windowClosed", testId, appControlChannel);
-    await wait(constants.WindowCloseWaitTime);
   };
 
   confirmAppNotFoundErrorReceived = (exception: DOMException) => {
@@ -85,69 +83,3 @@ export class OpenControl1_2 implements OpenControl<Context> {
     expect(context.context.type).to.eq(expectedContextType, openDocs);
   };
 }
-
-const waitForContext = (contextType: string, testId: string, channel?: Channel): Promise<Context> => {
-  let executionListener: Listener;
-  return new Promise<Context>(async (resolve) => {
-    console.log(Date.now() + ` Waiting for type: "${contextType}", on channel: "${channel.id}" in test: "${testId}"`);
-
-    const handler = (context: AppControlContext) => {
-      console.log(` waitforcontext hander reached`);
-      if (testId) {
-        console.log(` ${testId} VS ${context.testId}`);
-        if (testId == context.testId) {
-          console.log(Date.now() + ` Received ${contextType} for test: ${testId}`);
-          resolve(context);
-          if (executionListener) executionListener.unsubscribe();
-        } else {
-          console.warn(Date.now() + ` Ignoring "${contextType}" context due to mismatched testId (expected: "${testId}", got "${context.testId}")`);
-        }
-      } else {
-        console.log(Date.now() + ` Received (without testId) "${contextType}" for test: "${testId}"`);
-        resolve(context);
-        if (executionListener) executionListener.unsubscribe();
-      }
-    };
-
-    if (channel === undefined) {
-      console.log("adding system channel listener in waitForContext");
-      executionListener = fdc3.addContextListener(contextType, handler);
-    } else {
-      console.log(`adding app channel (${channel.id}) listener in waitforcontext`);
-      executionListener = channel.addContextListener(contextType, handler);
-      //App channels do not auto-broadcast current context when you start listening, so retrieve current context to avoid races
-      const ccHandler = async (context: AppControlContext) => {
-        if (context) {
-          if (testId) {
-            if (testId == context?.testId && context?.type == contextType) {
-              console.log(Date.now() + ` Received "${contextType}" (from current context) for test: "${testId}"`);
-              if (executionListener) executionListener.unsubscribe();
-              resolve(context);
-            } //do not warn as it will be ignoring mismatches which will be common
-            else {
-              console.log(
-                Date.now() +
-                  `Checking for current context of type "${contextType}" for test: "${testId}" Current context did ${context ? "" : "NOT "} exist,
-              had testId: "${context?.testId}" (${testId == context?.testId ? "did match" : "did NOT match"})
-              and type "${context?.type}" vs ${contextType} (${context?.type == contextType ? "did match" : "did NOT match"})`
-              );
-            }
-          } else {
-            console.log(Date.now() + ` Received "${contextType}" (from current context) for an unspecified test`);
-            if (executionListener) executionListener.unsubscribe();
-            resolve(context);
-          }
-        }
-      };
-      channel.getCurrentContext().then(ccHandler);
-    }
-  });
-};
-
-const broadcastCloseWindow = async (currentTest) => {
-  const appControlChannel = await fdc3.getOrCreateChannel(constants.ControlChannel);
-  appControlChannel.broadcast({
-    type: "closeWindow",
-    testId: currentTest,
-  } as AppControlContext);
-};
