+++ src/test/v2.0/support/intent-support-2.0.ts
@@ -3,7 +3,7 @@ import { AppIdentifier, Channel, IntentResolution, IntentResult, Listener, Priva
 import { Context, DesktopAgent, getOrCreateChannel } from "fdc3_2_0";
 import { APIDocumentation2_0 } from "../apiDocuments-2.0";
 import constants from "../../../constants";
-import { sleep, wait, wrapPromise } from "../../../utils";
+import { sleep, wrapPromise } from "../../../utils";
 import { AppControlContext, IntentUtilityContext } from "../../../context-types";
 
 declare let fdc3: DesktopAgent;
@@ -193,13 +193,6 @@ export class RaiseIntentControl2_0 {
   disconnectPrivateChannel(privateChannel: PrivateChannel): void {
     privateChannel.disconnect();
   }
-
-  async closeIntentAppWindow(testId) {
-    await broadcastCloseWindow(testId);
-    const appControlChannel = await fdc3.getOrCreateChannel("app-control");
-    await waitForContext("windowClosed", testId, appControlChannel);
-    await wait(constants.WindowCloseWaitTime);
-  }
 }
 
 const broadcastCloseWindow = async (currentTest) => {
@@ -208,60 +201,6 @@ const broadcastCloseWindow = async (currentTest) => {
     type: "closeWindow",
     testId: currentTest,
   } as AppControlContext);
-};
-
-const waitForContext = (contextType: string, testId: string, channel?: Channel): Promise<Context> => {
-  let executionListener: Listener;
-  return new Promise<Context>(async (resolve) => {
-    console.log(Date.now() + ` Waiting for type: "${contextType}", on channel: "${channel.id}" in test: "${testId}"`);
-
-    const handler = (context: AppControlContext) => {
-      if (testId) {
-        if (testId == context.testId) {
-          console.log(Date.now() + ` Received ${contextType} for test: ${testId}`);
-          resolve(context);
-          if (executionListener) executionListener.unsubscribe();
-        } else {
-          console.warn(Date.now() + ` Ignoring "${contextType}" context due to mismatched testId (expected: "${testId}", got "${context.testId}")`);
-        }
-      } else {
-        console.log(Date.now() + ` Received (without testId) "${contextType}" for test: "${testId}"`);
-        resolve(context);
-        if (executionListener) executionListener.unsubscribe();
-      }
-    };
-
-    if (channel === undefined) {
-      executionListener = await fdc3.addContextListener(contextType, handler);
-    } else {
-      executionListener = await channel.addContextListener(contextType, handler);
-      //App channels do not auto-broadcast current context when you start listening, so retrieve current context to avoid races
-      const ccHandler = async (context: AppControlContext) => {
-        if (context) {
-          if (testId) {
-            if (testId == context?.testId && context?.type == contextType) {
-              console.log(Date.now() + ` Received "${contextType}" (from current context) for test: "${testId}"`);
-              if (executionListener) executionListener.unsubscribe();
-              resolve(context);
-            } //do not warn as it will be ignoring mismatches which will be common
-            else {
-              console.log(
-                Date.now() +
-                  ` CHecking for current context of type "${contextType}" for test: "${testId}" Current context did ${context ? "" : "NOT "} exist, 
-had testId: "${context?.testId}" (${testId == context?.testId ? "did match" : "did NOT match"}) 
-and type "${context?.type}" (${context?.type == contextType ? "did match" : "did NOT match"})`
-              );
-            }
-          } else {
-            console.log(Date.now() + ` Received "${contextType}" (from current context) for an unspecified test`);
-            if (executionListener) executionListener.unsubscribe();
-            resolve(context);
-          }
-        }
-      };
-      channel.getCurrentContext().then(ccHandler);
-    }
-  });
 };
 
 export enum IntentResultType {
